guix: Add downloader for go modules from GOPROXY
Add a new downloader which implements, approximately, the download step of go
get $module. This is a convenient way of downloading zips with go modules by
just specifying the version and import path, as an alternative to git clone,
or awkward https:// fetches. This is particularly useful for sources that are
processed before release (like autotools generated files in tarballs) or
generated modules.

* guix/go-mod-download.scm: New file.
* Makefile.am (MODULES): Add it.

Change-Id: Ibb3b3ee70833fd0ea0c64278c95b8cb96a0be639