guix: Add module-aware build system for go
Add a new build system for go, using go modules. This build system is partly
compatible with go-build-system; they can both be used as build inputs to each
other, but their options are incompatible.

The main departure from go-build-system is that go-build-system tries to build
a workspace [1], where go-module-build-system builds a goproxy + go.mod file
and lets `go build` do what it wants to.

Most go libraries should be straight forward to build. For example, this is
the package definition for golang.org/x/sync@0.12:

(define-public go-golang-org-x-sync
  (package
    (name "go-golang-org-x-sync")
    (version "0.12.0")
    (source
     (origin
       (method go-mod-fetch)
       (uri (go-mod-reference
             (path "golang.org/x/sync")
             (version version)))
       (sha256
        (base32 "00pd84ah4xd5sjax8rxv98xbnwrvkk8clazl3kq1xrbkmvjq2m53"))))
    (build-system go-module-build-system)
    (home-page "https://golang.org/x/sync")
    (synopsis "Go Sync")
    (description "This repository provides Go concurrency primitives
in addition to the ones provided by the language and \"sync\" and
\"sync/atomic\" packages.")
    (license license:bsd-3)))

The build system also supports higher resolution build-, test-, and install
targets, re-use of compiled files, and options for common build tweaks.

[1] <https://golang.org/doc/code.html#Workspaces>

* guix/build/go-module-build-system.scm: New file.
* guix/build-system/go-module.scm: New file.
* Makefile.am (MODULES): Add them.

Change-Id: I47a028ab8f95fd3a338036480dbad6677e9c50a5