gnu: Add ansel.
* gnu/packages/photo.scm (ansel): A package definition for Ansel, a fork of
the Darktable. Separate package definition as the projects have deviated and
are continuing on their own trajectory. A git checkout is used as Ansel
only has nightly builds on version 0.0.0. The test code is only a
suggestion, and fails when running. The lack of working tests is described
in https://github.com/aurelienpierreeng/ansel/issues/208 . GCC 11 does not
work, so use GCC 12. Snippet removes submodules that are not needed.