gnu: Add unicorn-2.0.
This older version of unicorn is required by python-angr. Without
it, python-angr does not pass check [0]. angr upstream is aware of
the problem but does not yet support newer versions of unicorn [1].
Other distributions, like Nix, solve this by packaging the older
version of unicorn specifically for angr [2], let's do the same!

* gnu/packages/emulators.scm (unicorn-2.0): New variable.

[0]: https://issues.guix.gnu.org/78132
[1]: https://github.com/angr/angr/issues/4921
[2]: https://github.com/NixOS/nixpkgs/pull/388155