gnu: sane-backends-minimal: Respect backend search paths.
SANE's dynamic loading mechanism looks for configs in SANE_CONFIG_DIR,
and the libraries are expected to be either in LD_LIBRARY_PATH or in the
LIBDIR, which is set during the build to $PREFIX/lib/sane. A small patch
is necessary to make SANE respect backend search paths set by Guix.

* gnu/packages/scanner.scm (sane-backends-minimal/fixed): New variable to...
(sane-backends-minimal): ... replace this.

Change-Id: I132a709d171331d249acfe2fd1242dae43867ae8