gnu: services: base: Set a home directory for the "greeter" user
The home directory for the "greeter" user explicitly does not get created.
This causes an "unable to set working directory: no such file or directory (os
error 2)" error before every login prompt.  The reason for this is that greetd
starts agreety as the "greeter" user.  The working directory for the newly
created process is, by default, "/home/greeter", which does not exist (per our
instructions).  To fix this we set the home directory to "/var/empty".

While we are at it: set the login shell to "/sbin/nologin".

Change-Id: I6d971968e8b6a7825c261d4695caf0dd70127e5d