WIP: Package forgejo
This is the start of my attempt to package Forgejo.  gnu/packages/forgejo.scm
contains the results of repeated invocations of guix import.  These packages
should of course end up in the various golang-* modules and version-control
but I wanted to try to get it building before I started worrying about which
packages belong where.  Now I've run into a circular dependency issue that I
am not familiar enough with Go and its ecosystem to resolve.

Change-Id: I5400549e764ed65270f740b0e8e25d28132a0377