guix: node-build-system: Add temporary local NPM registry in the configure phase to work around an NPM CLI bug.
This change is a temporary fix to addresses NPM CLI bug https://github.com/npm/cli/issues/8364 which affects Guix's node-build-system.

The issue occurs when there is 2 or more dependencies, where one is either direct or indirect and the other(s) is/are indirect, and which have the same package name but have different versions. For example, this occurs for copy-props 2.0.5 which has is-plain-object 5.0.0 as a direct dependency, and is-plain-object 2.0.4 as an indirect dependency via each-props 1.3.2. Currenly with the version of NPM we have, is-plain-object 2.0.4 gets incorrectly deduped by the NPM CLI because it doesn't take the absolute path that we set as the version string into consideration. More details can be found in the GitHub bug.

This fix addresses the issue by creating a pseudo NPM registry locally that uses semantic version numbers instead of Guix store paths as the dependency version string:
  1. Resursively crawl all dependencies to get their versions and paths.
  2. Write metadata json files that mock the response from an NPM registry using the data we gathered in the previous step and using each package's package.json content where each dependency's store path string is transformed to the specific semantic version number of that package.
  3. Backup the current package's transformed package.json to be restored later and replace each dependency's store path to a semantic version number.
  4. Start a local http server to serve the metadata files.
  5. Run npm install with --registry=http://localhost:xxxx.
  6. Restore the package.json file with Guix store paths.