gnu: make-torbrowser: Add store to rdd allowlist.
Without this patch, some extensions break when the browsers are built with
system nss and livestreams are broken on these browsers due to a default
deny sandboxing rule.

See commit ab24e2ebe51720f332215b110c1bb151718d16bd and issues #546 on
codeberg and #77559 on debbugs.

* gnu/packages/tor-browsers.scm (make-torbrowser) [arguments] <#:phases>
'apply-guix-specific-patches: Add torbrowsers-add-store-to-rdd-allowlist.patch.
* gnu/packages/patches/torbrowsers-add-store-to-rdd-allowlist.patch: Add it.
* gnu/local.mk: Register it.

Change-Id: I4c9cb514131eae35db6149492e0ee1c4916bcfde