maths: hdf5: Comment on leaking internal dependencies zlib and libaec.
The patch 'hdf5-config-dependencies.patch' adds 'find_dependency' calls for
zlib and libaec in hdf5-config.cmake.  As a result, packages that use
find_package(HDF5) in CMake Config mode must add zlib and libaec to the build
environment.  This is a reminder that those additional inputs may no longer be
required after hdf5 is updated.

Change-Id: I0ad29d0379bea0cb99f1b5dd03868d14b3b551f6