guix: offload: Refactor check-machine-availability.
In order to conveniently use a self-loading machine detection script,
it's convenient to be able to check for a single machine.  This
refactor and export is a step in that direction.

This changes a bit the functionality of the consuming procedure
check-machines-availability-from-file: instead of performing each
check on all machines, it performs all checks on each machine.  It
doesn't affect correctness, and since this procedure is currently only
called in `guix offload test`, I didn't care to check for performance
changes.

* guix/offload.scm (check-machines-availability): Remove procedure, in
favor of...
  (check-machine-availability): Add procedure.