gnu: hdf5: Do not propagate libaec and zlib dependencies.
The patch 'hdf5-config-dependencies.patch' added 'find_dependency' calls for
libaec and zlib in hdf5-config.cmake.  As a result, packages that use
find_package(HDF5) in CMake Config mode must have libaec and zlib in the build
environment.  The patch appears to be unnecessary; it should be sufficient to
add libaec and zlib as inputs to packages that need them.

Related discussion in: <https://codeberg.org/guix/guix/pulls/953>.

* gnu/packages/maths.scm (hdf5): Do not propagate dependencies.
[source]<patches>: Remove hdf5-config-dependencies.patch.
* gnu/packages/image-processing.scm (insight-toolkit) [inputs]: Remove
libaec.
* gnu/packages/patches/hdf5-config-dependencies.patch: Delete file.
* gnu/local.mk (dist_patch_DATA): Deregister hdf5-config-dependencies.patch.

Change-Id: Idef61308cd146e898d1a4a8f1a97db23d36fc649