gnu: neovim: Add support for guix neovim packages
Enables the use of guix as a package manager for neovim, compatible with
the outputs of `vim-build-system`.

* gnu/packages/vim.scm: Added another phase to install the system init
file to the correct place.
* gnu/packages/aux-files/guix-nvim.vim: added installation of guix store
paths to neovim's runtimepath using lua, very similar to the current
guix.vim used for vim but installs to runtimepath instead of
packagepath in order to facilitate packages that cannot be installed to
the packagepath.

Change-Id: I375e3e0d051da27d065dff5456cb298a01bc90b1