gnu: neovim: Add guix.vim equivalent aux file
Enables the use of guix as a package manager for neovim, compatible with
the outputs of `vim-build-system`.

* gnu/packages/vim.scm: Changed native-inputs to old style, and added
`guix-nvim.vim`. Then added another phase to install the system init
file to the correct place. Restyled with `guix style`.
* gnu/packages/aux-files/guix-nvim.vim: added a copy of
`gnu/packages/aux-files/guix.vim` but with `/share/nvim/site` instead of
`/share/vim/vimfiles`.

Change-Id: I375e3e0d051da27d065dff5456cb298a01bc90b1