build-system/guile: Rename %scheme-file-regexp to default-scheme-file-regexp and export.
In 0559a4b547 default values have been removed from parameters to the
build phase of the guile build system. This broke the package
"libfive" which calls that phase directly. One of those para,egter
"%scheme-file-regexp" was a define within the guile build system. To
avoid duplication, this commit exports this variable so it can be used
within the package definition of "libfive". To signify that it is no
longer an internal variable, it was renamed from "&scheme-file-regexp"
to "default-scheme-file-regexp".

* guix/build-system/guile.scm (%scheme-file-regexp): Delete variable,
rename to default-scheme-file-regexp.
(default-scheme-file-regexp): New variable, renamed from
%scheme-file-regexp. Export.

Change-Id: Ia4cfe59b4d0e9fda9555268f5c22e153123fc2c6