gnu: Add insight-toolkit-legacy.
This is insight-toolkit (ITK) built with ITK_LEGACY_REMOVE=OFF.  With this
option (which is the default), ITK is built including API marked for removal
in the next major version.  Note that enabling Python wrapping forces this
option to ON, so Python wrapping is not enabled for this build, which is
another difference.  ITK-SNAP 4.2.2 requires ITK >= 5.4 built in this way.

* gnu/packages/image-processing.scm (insight-toolkit-legacy): New variable.

Change-Id: Ief18c1b4241241881b5b632121fc463db5b3c286