guix: packages: Make git-version a record.
The goal of this in the long term is to move from
```
(let ((commit "...")
      (revision "0"))
  (package
    (name "...")
    (version (git-version "0.0.0" revision commit))
    (source
     (origin
       (method git-fetch)
       (uri (git-reference
              (url "https://...")
              (commit commit)))))
    ...))
```

to

```
(package
  (name "...")
  (version (git-version (semantic "0.0.0")
                        (revision "0")
                        (commit "...")))
  (source
   (origin
     (method git-fetch)
     (uri (git-reference
            (url "https://...")
            (commit (git-version-commit version))))))
  ...)
```

In the meantime, this implementation proposes 3 replacements for
git-version:
1. Drop-in `vc-version`: Exact same function, renamed. It
should only be used when we're not using `git-fetch`.
2. Drop-in `make-git-version`: Exact same arguments, but returns a
`<git-version>` instance.
3. `<git-version>` instance, used as above, except during the
transition, we need an intermediary name, here `git-version*`.

Warning: `#$version` in a G-expression (resp. `,version` in a quoted
S-expression, or similarly `(string-append things ... version)`)
doesn't have the same meaning when it's not a string. The drop-in
replacement, a bit less convenient is `#$(package-version
this-package)`.

* guix/packages.scm
(<package>)[version]: Extend to <git-version> instance.
(package->bag): Ensure the exact same string version is passed to
bags.
(<git-version>): New record.
(package-version): Change procedure, to always return a string
version (and not a <git-version>).
(warn-unexpected-package-version): Add smart location-aware error
procedure.

* guix/git-download.scm
(vc-version): New procedure, pure copy from former git-version.
(git-version): Deprecate, and use vc-version under the hood.
(git-file-name): Handle <git-version> case.