gnu: openscad: Update to 2025.07.25
OpenSCAD is designed to be built with the build dir being a direct
descendant of the source dir. As it was becoming more and more hacky
to try to force an out-of-source build like it's common with Guix,
this update changes the build dir to be a direct descendant of the
source dir, as expected by OpenSCAD, to avoid issues with test cases,
which would have needed more and more quirky workarounds with each new
version to accomodate new tests.
The project originally uses CMake for building and since using the
qt-build-system caused issues with the latest version, the build
system is being switched to CMake as well.

* gnu/packages/engineering.scm (openscad): Update to 2025.07.21,
  switch to under-source-build, switch to cmake-build-system
  [source]: Remove patch
  "openscad-fix-path-in-expected-test-results-to-acommodate-diff.patch"
  as it was only used for out-of-source builds.
  [build-system]: Switch from qt-build-system to cmake-build-system.
  [arguments]<#:out-of-source?>: Do not perform an out-of-source
  build.
  [arguments]<#:configure-flags>: Set the build dir to be a descendant
  of the source dir.
  [arguments]<#:phases>[patch-source]: Remove substitute* for
  "tests/test_cmdline_tool.py" as this was only needed for
  out-of-source builds.
  [arguments]<#:phases>[create-build-dir]: Add phase.
  [arguments}<#:phases>[patch-tests]: Do not copy color-shemes and
  shaders to the build directory as this was only necessary for
  out-of-source builds.

Change-Id: I3f073ea3a6a816b2cee5d271e9038721300deb70