daemon: handle sockets in secureFilePerms
When using the -K argument to guix build, before publicly exposing the build
directory, we check that all the files we expose are "safe"; mainly that they
aren't world-writable and don't have setuid/setgid bits. If there is a file
type we don't understand, we bail.

Some packages create sockets during their build process, which currently
prevents -K from working because sockets are not one of the file types it
handles. For example, if plasma-desktop is failing to build (as it is
currently for me on aarch64 due to an unrelated issue), the build command with
-K will throw an exception and not finish exposing the
directory. setuid/setgid is meaningless for sockets so we just clear the
world-writable flag.

This change extends the logic to handle sockets and now guix build -K
plasma-desktop works fine.

* nix/libstore/build.cc (secureFilePerms): handle files of socket type.

Change-Id: I129be06285686ed34af0a25c373896bcb0bf0ebd