gnu: mono: Fix CFLAGS for  mono@3.12.1.
During a recent update, the styling of the package was adadpted to follow the guix style guides.
During this change, it was forgotten to add spaces in between the CFLAGS which causes./configure to
fail when testing the compiler.
This patch addes the missing spaces, closely following the styling of mono@2.4.2.

* gnu/packages/dotnet.scm (mono-3.12.1)[arguments]<#:phases>[set-cflags]: Add spaces in between
CFLAGS.

Change-Id: Ia15ed59bd9a69a97c3fe722ece862bcb065d2abf