gnu: Remove python-pafy.
This package is abandoned / unmaintained upstream, the original source repo
has been deleted, and it no longer works as expected.  While yt-dlp is
command-line-compatible with youtube-dl, it’s not source-compatible, which is
how this package uses it.  Since 753588606c9ca0acdab15bc828f8f33b7bcf0172
replaced youtube-dl with yt-dlp, this package hasn’t worked as expected, and
it prints a warning when building Guix:

    warning: 'youtube-dl' is deprecated, use 'yt-dlp' instead

No other packages in Guix depends on this; remove it.

* gnu/packages/python-xyz.scm (python-pafy): Delete variable.

Change-Id: If893446a3827a0d513ab3cb8a9c6148b2b062f24