gnu: ghc-9.4: Fix build for bootstrapping GHC 9.8's Hadrian
The testsuite driver's worker thread pool gets filled with zombie
processes which causes it to hang and the build to time out. No idea
what might be causing that, but as 9.4 is only used to build later GHC
versions it should be enough that they pass their testsuites.

* gnu/packages/haskell.scm
(ghc-9.4): Downgrade to 9.4.3 for bootstrapping GHC 9.8's hadrian.
<arguments>[#:tests?]: Set to #f to disable failing testsuite.

Change-Id: I2c889f5fc7a6898562fcb8377fc1f582ad835434